data_quality_audit/pyproject.toml
[project]
name = "data-quality-audit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.59",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
]
